pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eoapi-notifier"
version = "0.0.3"
description = "Message handler for eoAPI components"
requires-python = ">=3.12"
authors = [
    { name = "Development Seed" }
]
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "PyYAML>=6.0.2",
    "click>=8.2.1",
    "asyncpg>=0.30.0",
    "paho-mqtt>=2.1.0",
    "cloudevents>=1.10.0",
    "httpx>=0.25.0",
    "pypgstac[psycopg]>=0.9.8",
]
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.scripts]
eoapi-notifier = "eoapi_notifier.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=8.4.2",
    "pytest-cov",
    "pytest-asyncio",
]
dev = [
    "pre-commit",
    "mypy",
    "ruff",
    "types-PyYAML",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["UP046"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"integration/*" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "integration.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false

[tool.pytest.ini_options]
addopts = [
    "--cov=eoapi_notifier",
    "--cov-fail-under=80",
    "--cov-report=term-missing",
    "--cov-report=xml",
]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["eoapi_notifier"]
omit = ["tests/*", "eoapi_notifier/logging.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "logger\\.debug",
    "logger\\.info",
    "logger\\.warning",
    "logger\\.error",
    "logger\\.critical",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250822",
]
